# [5.5] Respect abstract classes in kernel->load()