If you happen to have an `abstract class BaseCommand extends \Illuminate\Console\Command` inside `app/Console/Commands` folder the whole cli app is broken.

```
$ php artisan
  [Illuminate\Contracts\Container\BindingResolutionException]
  Target [App\Console\Commands\BaseCommand] is not instantiable.
```

This PR fixes this issue.

Note: there is a more strict way to perform a check - http://php.net/manual/en/reflectionclass.isinstantiable.php but it looks like overkill to me