@devcircus 2017-08-24
Maybe we should just disable the auto register if we're using an abstract command. Adding reflection here seems like too much.