@taylorotwell 2017-08-25
I would maybe just try / catch and call `report` if it fails.