@themsaid 2017-08-25
Yes I really think we should use Reflection as it's too expensive, for an app with lots of commands it could affect bootup cost running every single console command.