@arrilot 2017-08-25
Okay, I've changed it to `try / catch / report` following a Taylor's suggestion.