@davejamesmiller 2017-08-26
This does make it work... but the downside of reporting the error is the log quickly starts to fill up with errors that aren't really errors:

```
[2017-08-26 16:03:39] local.ERROR: Target [App\Console\Commands\Download\DownloadCommand] is not instantiable. {"exception":"[object] (Illuminate\\Contracts\\Container\\BindingResolutionException(code: 0): Target [App\\Console\\Commands\\Download\\DownloadCommand] is not instantiable. at .../vendor/laravel/framework/src/Illuminate/Container/Container.php:918)
[stacktrace]
... 11 more lines of stack trace ...
```