@davejamesmiller 2017-08-26
> Yes I really think we shouldn't use Reflection as it's too expensive, for an app with lots of commands it could affect bootup cost running every single console command.

I've been doing some tests to see how expensive it is. On my machine, if you have at least one abstract class, it's actually *slower* to catch and log the error than it is to check them all with `isInstantiable()`. Plus, as I mentioned above, it does fill up the logs.

On the other hand, if you don't have any abstract classes (or you don't log the errors), using `isInstantiable()` every time would add a small amount of load time (see below for measurements).

But what if we combined the two and check if it's abstract *inside* the catch? Then we don't incur the performance penalty for every class. If it is abstract, we can skip it quietly, because it's obviously not a valid command. If it's not abstract, we can either rethrow the error (my preference, so the user sees the error on screen) or log it (as in the PR).

This is my proof of concept code - it can be tidied up to make it more readable if accepted.

```php
if (is_subclass_of($command, Command::class)) {
    Artisan::starting(function ($artisan) use ($command) {
        try {
            $artisan->resolve($command);
        } catch (ContainerExceptionInterface $e) {
            if ((new \ReflectionClass($command))->isInstantiable()) {
                throw $e;
            }
        }
    });
}
```

Thanks!

**Details of time measurements:**

My app has 31 commands + 1 abstract class. I'm using a Linux VM with a hard disk (not SSD). I used the `time` command (`real` measurement) to run `php artisan` with no parameters, repeated it several times in a row and recorded the range I observed, excluding outliers. So not entirely scientific, but I think accurate enough.

- Catch and log the error (current PR) - 0.29s to 0.34s
- Catch with no reporting - 0.24s to 0.26s
- Using ReflectionClass + isInstantiable() - 0.25s to 0.27s
- Using reflection inside catch (my code above) - 0.24s to 0.26s
- Using load() with no abstract classes - 0.24s to 0.26s 
- Not using load() at all (using $commands instead - 0.24s to 0.26s

Separately I used microtime() to measure how long isInstantiable() actually takes to run on all 32 classes - it was about 0.011s, which is consistent.

(Incidentally, the Container already does the isInstantiable() check for every class - but because it throws the same BindingResolutionException for any error, the only way to differentiate it from other errors, as far as I can see, would be (a) string matching on the message (ugh) or (b) add a new subclass to Container (overkill).)