@taylorotwell 2017-08-30
Yeah, it may be fine to use reflection then.