@arrilot 2017-08-30
Thanks for review @davejamesmiller 
I've tweaked some code once again and changed the base branch from `master` to `5.5`
Should be ready to be merged now.

@taylorotwell Is there anything else you want to be changed here?